+++ lib/src/app/build.dart
@@ -2,15 +2,21 @@ part of layrz_models;
 
 @freezed
 class AppBuild with _$AppBuild {
+  /// [AppBuild] is the build of the app
   const factory AppBuild({
+    /// [id] is the ID of the build
     required String id,
+
+    /// [buildName] is the semantic name of the build
     required String buildName,
+
+    /// [buildNumber] is the semantic number of the build
     required int buildNumber,
+
+    /// [fileUri] is the URI of the file, only used when the platform is distinct to `AppPlatform.web`
     String? fileUri,
-    String? comments,
-    @AppBuildStatusConverter() required AppBuildStatus status,
-    @TimestampOrNullConverter() DateTime? startAt,
-    @TimestampOrNullConverter() DateTime? endAt,
+
+    /// [isNext] is the flag to indicate if the build is the indev environment
     bool? isNext,
   }) = _AppBuild;